Remove use of legacy context
